// my-openapi-zod-app/src/api/GreenOnionClient/Schema/ZQueryIngredientModel.ts
//
// This file is generated. Do not edit manually.
//
// @applicationName: GreenOnion.API
// @applicationDescription: GreenOnion.API - 1.0.0.0
// @applicationVersion: 1.0.0.0
// @generatedDate: 2025/07/25
//

import { z } from "zod";

export const ZQueryIngredientModel = z.object({
    ingredientId:z.number().int().optional(),
    ingredientName:z.string().optional(),
    group:z.string().optional(),
    iocGroup:z.string().optional(),
});